fix: should use float64 instead of int (so weird)

+++ StatisticsApi.go
@@ -4,13 +4,13 @@ import "github.com/sirupsen/logrus"
 
 type FileStatistics struct {
 	// 文档历史访问人数，同一用户（user_id）多次访问按一次计算。
-	Uv int `json:"uv"`
+	Uv float64 `json:"uv"`
 	// 文档历史访问次数，同一用户（user_id）多次访问按多次计算。（注：同一用户相邻两次访问间隔在半小时内视为一次访问）
-	Pv int `json:"pv"`
+	Pv float64 `json:"pv"`
 	// 文档历史点赞总数，若对应的文档类型不支持点赞，返回 -1
-	LikeCount int `json:"like_count"`
+	LikeCount float64 `json:"like_count"`
 	// 时间戳（秒）
-	Timestamp int `json:"timestamp"`
+	Timestamp float64 `json:"timestamp"`
 }
 
 func (c AppClient) NewStatistics(data map[string]any) *FileStatistics {
@@ -18,10 +18,10 @@ func (c AppClient) NewStatistics(data map[string]any) *FileStatistics {
 	// 其中statistics是一个map[string]any
 	statisticsMap := data["statistics"].(map[string]any)
 	return &FileStatistics{
-		Uv:        statisticsMap["uv"].(int),
-		Pv:        statisticsMap["pv"].(int),
-		LikeCount: statisticsMap["like_count"].(int),
-		Timestamp: statisticsMap["timestamp"].(int),
+		Uv:        statisticsMap["uv"].(float64),
+		Pv:        statisticsMap["pv"].(float64),
+		LikeCount: statisticsMap["like_count"].(float64),
+		Timestamp: statisticsMap["timestamp"].(float64),
 	}
 }
